test:use testdir (#18)
* Use testdir at async test to allow parallel execution

testdir()! call creates test folders inside a predetermined path. In my
machine that is:

`/var/folders/v4/x_j3jj7d6ql4gjdf7b7jvjhm0000gn/T/testdir-of-stelios/`.

The content of this dir is:

```
.
├── tofnd-0
│   ├── cargo-pid
│   └── tofnd
│       ├── kv_manager
│       │   └── tests
│       └── tests
│           ├── basic_keygen_and_sign
│           └── restart_one_party
├── tofnd-1
│   ├── cargo-pid
│   └── tofnd
│       ├── kv_manager
│       │   └── tests
│       └── tests
│           ├── basic_keygen_and_sign
│           └── restart_one_party
└── tofnd-current ->
/var/folders/v4/x_j3jj7d6ql4gjdf7b7jvjhm0000gn/T/testdir-of-stelios/tofnd-1
```

Everytime a new test is executed, a new `tofnd-X` folder is created. The
purpose of this is to observe the results of each test.

* Use testdir at kv_manager unit tests

* Satisfy cargo-fmt and clippy

* Use reference instead of clone()

* Move kv_manager into a directory and decouple tests